Fix docker build (#1016)
* Improves Docker setup for production builds

Refactors the Dockerfile to use multi-stage builds, optimizing for production environments.

Introduces a builder stage to compile the application and a runner stage for serving the built assets. This reduces the final image size and improves security by including only the necessary files.

Configures Next.js to produce a standalone output, minimizing dependencies and improving deployment consistency.

* Adds Docker build check to CI

Adds a job to the CI workflow that checks if the Docker image builds correctly.
This ensures that the Dockerfile is valid and the build process is working as expected before merging changes.
The image is not pushed to any registry and is not used anywhere, as the final image is built in the GitLab CI pipeline.